move: auth 폴더 위치 원래 위치로 이동 (#124)
