+++ apps/web/src/app/(with-global-layout)/auth/callback/page.tsx
@@ -1,4 +1,4 @@
-import AuthCallbackPage from 'src/features/auth/AuthCallbackPage';
+import AuthCallbackPage from 'src/features/user/auth/AuthCallbackPage';
 
 const AuthCallback = () => {
   return (
