+++ apps/web/src/app/(with-global-layout)/auth/signup/page.tsx
@@ -1,4 +1,4 @@
-import SignUpPage from 'src/features/auth/SignUpPage';
+import SignUpPage from 'src/features/user/auth/SignUpPage';
 
 const SignUp = () => {
   return (
